Erro de peso

+++ src/main/java/br/wk/projeto/bancosangue/dto/DoadorDTO.java
@@ -50,5 +50,5 @@ public class DoadorDTO {
     @JsonProperty("altura")
     private Double altura;
     @JsonProperty("peso")
-    private Double peso;
+    private int peso;
 }
